Backend/ocr-service/requirements.txt
Flask==2.1.0
pytesseract==0.3.8
Pillow==8.3.2
pdf2image==1.16.0
PyPDF2==2.0.0